Add scripts/update_siggen.py
This adds a script that updates socorro-siggen which is an extracted library
for socorro signature generation.

It's not perfect. Namely it's possible for us to land changes in socorro
signature generation that won't work in siggen. So we'll have to do some
back and forth. It does make that a lot easier.