fix bug 1473068: fix session refresh errors with xhr
The search page does XHR requests when switching between searches. Other
pages do similar things. When the SessionRefresh middleware handles those
requests, it can return a 403 or a redirect, but the js can't handle that.

This fixes the SessionRefresh middleware to ignore those urls and not
refresh the session on them.