fix bug 1477013: rewrite collapse to handle rust trait methods
When signature generation normalizes frames, it takes the function and
collapses it down to a more "neutral form" which creates a signature that
is more likely to match other crash reports that have the "same problem".

That works fine with functions from C/C++ but doesn't work well with
functions from Rust where < and > has different meaning.

This reworks the collapse code to simplify it a bit, adds tests, adds
documentation, and adds handling for the case where the stuff between
< and > is a Rust trait method.