fix bug 1483097: fix product home when there's no build/version info
This fixes the product home when there's no information for that product
in product_versions. When that happens, build_default_context will pull
the version data from super search.

This also fixes the navigation bar to display products in context["products"]
so all products will show up regardless of whether they have version
info or not.