bug 1481696: add processor rule to fix Focus crashes
This reworks the ProductRewrite rule so it logs what it did in the crash
report better. It also adjusts the rule to act as a "this is where we're
going to do all the product rewriting" place.

This adds product rewriting logic for Focus GV. That's using GeckoView,
but incoming crash reports are using ProductName=Fennec. So the ProductRewrite
rule was convering those to FennecAndroid. This will then convert it to
Focus if the ProcessType=content.

We should remove this as soon as we can, but it'll fix an existing issue
we have right now.