Switch transaction executor to use reraise
replace uses of raise in the database/transaction_executor class with six.reraise(). This revealed a weird line where raise was used outside of an exception block. In order for the program to get to that point, an error must have been silenced. This patch remedies the issue by keeping track of the sys.exc_info of the last error to use outside of the except block context.