Remove scrubber code
First off, there are two sets of things called "scrubber". One thing is part
of the Cleaner which goes through a Python data structure and removes items
that aren't in a whitelist (which should get renamed). We're keeping that.

The other thing takes text and "scrubs" URLs and email addresses from it.
Socorro used that solely to remove PII from user comments which are stored
in the raw crash "Comments" field and the processed crash "user_comments"
field. Since we changed user comments to be considered PII, we no longer
need this scrubber. This commit removes it.