fix bug 1490465: implement session_with_retries
We have a whole bunch of code that uses the Python requests library to
fetch data from various places. Some of that code has additional handling
for HTTP 429 responses. Some of it has connection timeouts. None of that
code sets the user agent to something helpful.

This fixes that by creating a function that generates a requests Session
instance with an adapter that sets a helpful user agent, sets a default
timeout, and retries on 429s and 500s. This updates the code to use
this new session. This also updates a bunch of tests to use requests_mock.