Implement session_with_retries
We have a whole bunch of code that uses the Python requests library to
fetch data from various places. Some of that code has additional handling
for HTTP 429 responses. None of that code sets the user agent to something
helpful.

This fixes that by creating a functino that generates a requests Session
instance that has a better user agent and handles 429s. It also updates
the myriad of requests-using places to use it.