fix bug 1323033: add doc_values=True to ES mapping
This adds doc_values=True to fields that are either not strings or explicitly
have index=not_analyzed. Per the ES docs, this will cause ES to store data
on disk for sorting and aggregations rather than in memory which should
reduce some of our memory problems when doing aggregations.