fix bug 1323033: fix doc_values for nested and boolean cases
This fixes doc_values in Elasticsearch mappings for the boolean and nested
cases. boolean fields can't have doc_values=True. This isn't documented
anywhere, but appears to be the case as near as we can determine.

This also fixes nested cases where a field may have sub fields or be
of an object type with properties. The code now traverses the structures
adding doc_values=True where appropriate.