fix bug 1478080: vendor crontabber
This brings the necessary bits of crontabber back into Socorro and updates
it to pass linting and work with Python 3.

This doesn't bring in the tests since those are written for nose and there's
a lot of it and we really just want to test Socorro crontabber jobs--not an
entire library. If it turns out that was a bad decision, we can add (and
overhaul) them later.