fix bug 1493200: fix empty string signature generation
There were two situations and a bug that were causing signature
generation to generate signatures of an empty string.

The bug was stomping on the signature regardless of what CSignatureTool
figured out. We only want to stomp on the signature if we have a valid
signature. That's fixed in this commit.

The two situations that weren't handled correctly both involve a space
between the last token and the previous token where the two tokens
together are the function name, args, and some [clone .cold.xyz] suffix.
This fixes that by making sure we're not losing bits.