fix bug 1457747: move graphics_device to Django
This creates a new GraphicsDevice Django model, rewrites all the bits to use
that, removes the old GraphicsDevices pseudo-middleware, and creates a data
migration to move the data from the old table to the new table.

Now the webapp owns knowledge and maintenance of graphics devices.

The thing this doesn't do is delete the old table--we'll do that in a later
step once we know everything is good.