Fix table and migration after testing with prod data
The data in prod is ... interesting.

I don't know whether to torch it or not, so for now I tweaked the table and
data migration to ditch the obviously bad data--anything that doesn't
have at least a vendor_name or an adapter_name. I figure if it's missing
both of those, then it's not doing anything, so we should ditch it.
That halves the data set and gets rid of most of the egregious junk.

We can figure out better data validation later.