Clean up transaction_executor
When we vendored crontabber, we added a clause to the exception handling
in the transaction_executor which crontabber had in its version. This
fixes that code so it's not importing the postgres code. That way
it's generalized again. This also moves the code from socorro.database
to socorro.lib since it's used in all kinds of places most of which
aren't databasey.

Additionally, this removes code and tests for the abandon_transaction
feature that isn't used.