bug 1493768: create BugAssociation Django model
This creates a BugAssociation Django model and updates the bugzilla
cron job to save data to both the old and new tables.

This also adds a data migration script.

This also removes test_run_job_based_on_last_success. After we removed
the bugs table a couple of years ago, this test was adjusted and the
bugs table bits removed. As near as I can tell, this doesn't test anything
useful anymore.