bug 1493768: redo webapp to use BugAssociation Django model
This redoes the webapp to use the BugAssociation Django model instead of
the old Bugs middleware model.

There are a few things to point out.

First, the Bugs middleware model did this funny thing where if you passed
in signatures, then it would find all the bugs associated with those
signatures and then find all the signatures associated with all those bugs.
I redid this as .get_bugs_and_related_bugs().

Second, some of the code that used the old Bugs middleware model relied on
this behavior. I redid that using .get_bugs_and_related_bugs(). Some of
the code filtered the results down to just the (bug, signature) sets for
signatures it asked for. I redid that code using just .filter().

Third, I changed the shape of things, so I had to change templates to
look at the bug_id field instead of the id field.

This does not remove the old table, but it does remove all use of it.