bug 1498441: rewrite everything to use new Platform model
This rewrites everything to use the new Platform Django model.

This also removes the /api/Platforms/ endpoint. I checked the dashboards
and it's not being used. It returned a static result, so anyone can just
hardcode that.

This removes the staticdata for the OsName model. In its place, it adds
a Django fixture and adds a line to the setup script to load that.

This also adds a schema migration to fix the crashstats_platform table
so that "name" is unique.