fix bug 1501780: rewrite BetaVersionRule to use Buildhub
This removes the VersionString API which we no longer need since the
BetaVersionRule was the only user of it.

This removes the AuroraVersionFixitRule because we don't need it anymore.
That rule was in place because ftpscraper and the related stored procedures
didn't handle devedition aurora well causing the BetaVersionRule to muck
up version strings for b1 and b2. Buildhub doesn't have that problem so
we can get rid of it.