bug 1493687: fix /crontabber-state
This nixes the CrontabberState middleware in favor of one that uses the
Django model.

This also removes the dependency visualization on the /crontabber-state
page. It doesn't like it when there are no dependencies and I figured
it's time to remove it.

Since it's the only use of d3-sankey, I removed that, too.

One minor irritant is that the depends_on field was a text[] in the
old models and a text (comma-separated string) in the new one. That's
causing some problems, so I hard-coded things. In the rewrite we won't
be redoing the depends_on, so it's not long for this world anyhow.