fix bug 1501780: fix b99 and reduce buildhub hits
ftpscraper would generate a product_versions record for X.0b99 which is a
"final beta". This fixes the BetaVersionRule to do a lookup for
(product, build id, beta) channel and if that returns nothing and the
version is possibly a final beta version, it will do a second lookup for
(product, build id, release) and if that returns something, it adds a b99
to the end of the version string.

Buildhub only stores build information for Firefox and Fennec, so we shouldn't
bother hitting it for other products. This fixes that.

This also splits up the monolithic "test_everything" into seprate tests.