fix bug 1503034: change supersearch form to pull versions from supersearch
The supersearch form suggests and autocompletes versions in the version input
field. The list of versions used to come from the ProductVersionMiddleware
which pulled versions from the product_versions table for the Firefox product.

This reworks that to pull versions from a facet of crash reports for Firefox
over the last 3 months and no longer rely on the product_versions table.