fix bug 1504040: redo context to use Supersearch data
This redoes the context to use supersearch data for all products
rather than just the products that don't have product_versions information.

This adds a ProductVersionsMixin for setting up the mock for setting
product versions in the tests.

This moves BaseTestViews into conftest.py. Seems like a better place
for scaffolding that a bunch of things use.