fix bug 1505233: gut Rule and simplify it
Previously, Rule was implemented so that a rule could stop processing for
a crash report. We don't use that feature, so we can nix that and stop
checking the return of action and act.

Rule had action/_action and predicate/_predicate where action and predicate
had a bunch of error handling. Instead of handling errors in Rule, we're
going to let them get handled by the thing running the rules. Rules should
be defensive and thoughtful in how they execute and not throw errors
willy-nilly.

Rules used to have versions. I took all that out since it's not used anywhere.

After simplifying Rule, there wasn't anything to test, so I removed the
Rule tests.

There were some other subclasses of Rule, but those weren't used, so I
removed those and their tests.