bug 1506907: implement Product model
This creates a Django model for products, creates a data migration to move
product data from the old table to the new one, and reimplements things to
use the new table.

We have to leave the old table in place because it's a foreign key of
the old product_versions table and some other tables. All that will go
away in a future commit.