bug 1507916: fix archive scraper scraping
Previously, archivescraper would only grab build information for a single
platform per buildN/ directory. That turns out to be wrong, so this fixes
it to get build information for ALL platforms per buildN/ directory.

However, that radically increases the runtime of archivescraper, to this
adds an optimization where it looks at the db for the maximum major version
for that product and only scrapes information for versions greater than
(major_version - 3). That significantly reduces the run time.