Remove dead code and rework django tables handling
The socorro tests (not the webapp tests) run without Django initializing
the database. Previously, there would be a bunch of code to create
the tables and maintain them, but this was spread out across the tests.
At this point, it seems prudent to pull it all together into a fixture.

However, a bunch of the tests derive from unittest.TestCase, so this
creates a mixin as well as a fixture to cover all the bases.

This also tweaks some tests so they work with the new system.