fix bug 1503434: redo retries and transactions
Socorro has a family of TransactionExecutor classes which perform transaction
bookkeeping and also retry failed operations. The way it's implemented doesn't
work with Python 3. Rather than fix the issues, this splits the
TransactionExecutor into a retry function and a transaction_context context
manager and then redoes everything to use those.

This does a couple of things:

1. simplifies executing operations with transaction bookkeeping requirements
   so they're easier to reason about

2. simplifies retriable operations so they're easier to reason about

3. gets rid of one or both in places that don't need them