bug 1515805: fix breakpad client building in image; add scripts
Building the breakpad client requires using depot_tools which only work with
Python 2. This fixes the build-breakpad.sh script so it makes sure depot_tools
are using Python 2 in the Docker container.

This adds a scripts/run_mdsw.sh script which runs minidump-stackwalker just
like we do in the processor, but outside of the processor. This makes it
a lot easier to debug issues with minidump-stackwalker.