Make the `app` user own the contents of /app in the Docker container.
This will allow the `app` user to build Breakpad (which requires creating
directories within /app) outside of docker-compose, which has different
permissions for /app since it mounts a local directory there.

Since the current pre-built Breakpad is broken, this also disables downloading
it pre-built so that we can build and push this image to DockerHub for use in
rebuilding the pre-built Breakpad.