fix bug 1519802, bug 1305956: fix cpu_info, cpu_arch/cpu_name, and cpu_count
Previously, "cpu_info" was composed of "cpu_info" and "cpu_count" from the
system_info in the json_dump. Socorro indexed this and also cpu_count
so they were both available for searching. The report view showed the
cpu_info value.

But this is confusing in a few scenarios and there doesn't seem to be a good
reason to have a contrived value like this.

This changes it so cpu_info is just the cpu_info value from
json_dump.system_info. It also adds cpu_count to the crash report and the
report view.

This also changes the name from "Build Architecture Info" (which doesn't
make sense) to "CPU Info".

While doing that, I removed two deprecated items from Elasticsearch fields
and also moved the whole cpu_name -> cpu_arch deprecation another step.