fix bug 1520570: switch to multi-stage docker image builds
This moves building breakpad client into a separate docker image build stage.
That way all its intermediary build tools and files don't swamp the final
socorro_app image and make it super huge.

This also adds support for "NOCACHE=1 make build" which will build the
docker image with --no-cache flag. I do that so often that it seemed
prudent to codify it in the Makefile and save myself some time.

This also tweaks the .dockerignore file to ignore breakpad client
build bits during the COPY which might show up in a local dev environment
when working on updating the breakpad client.