fix bug 1315109: don't show correlations tab if not supported
This moves the list of supported products for correlations to the view
code and fixes the JS to build the options based on that. The correlations.js
code still has its own list of supported products, but we can fix that
another day.

This also adjusts the jinja template such that if the specified product
isn't supported by correlations, it won't show the correlations tab at
all.