fix bug 606630: show friendly build date
This fixes the display of the build id in the report view so that it shows
the build id, a humanized version of the build id date, and then an explicit
link to the Buildhub data for that build.

This also adds the channel to the Buildhub link--that will focus the data
better. That'll help a little with bug #1056217.

This also moves the buildhub base url to settings--we shouldn't have had
it hardcoded in the template.