fix bug 1438001: redo date/time inputs sans flatpicker
This nixes flatpicker usage and replaces it with HTML date/time input
tags. This fixes problems with local date/times in the UI and
UTC being used in the API and data.

This also adds a "day ends" shortcut for setting the from/to times
to beginning and end-of-day. That's something I've wanted for a while.