fix bug 1419899: implement __heartbeat__ and __lbheartbeat__
This implements the __heartbeat__ and __lbheartbeat__ dockerflow endpoints
for the webapp. It reworks the existing healthcheck code and deprecates
the old version. We can remove that as soon as we update the infrastructure.

This also moves __version__ from the crashstats app to the monitoring
app; seems to be a more appropriate place.