bug 1527728: fix versions in nav bar
The versions in the nav bar are pulled from Super Search, so they're based
on the crash reports we've seen for that product in the last 60 days.

This fixes the query so that it's not cached in the SuperSearchUnredacted model
because it's cached in the get_versions_for_product. This also increases the
_facets_size so that it picks up versions which have recently been released
for which we don't have many crash reports, yet.