bug 1493687: write a Django cron command
This reimplements the bulk of crontabber as a Django cron command. It
continues to support retrying and backfill and frequencies and time-of-day.

Cron jobs are implemented as Django commands. Any Django command can be
a cron job regardless of whether it was implemented as a cron job originally
or not.

All the scheduling, locking, and other bookkeeping work is done by the
"cron" command.