bug 1527345: reduce Pub/Sub pulling
This reduces the number of Pub/Sub pulling operations by changing it to pull up
to 5 messages at a time from each queue. Further, it reduces the number of
times it pulls from the priority queue such that it only pulls from the
priority queue once per pass rather than twice per pass.

The theory is that this will increase the throughput for crash processing
on the standard queue which is where the vast majority of our messages
are.