bug 1539233: check if missing crashes have since been processed
This adds a check to verifyprocessed crontabber job to first check
to see if any of the crashes in the missing crashes table have since
been processed.

This also adds a filter to the Django admin to make it easier to
distinguish missing and non-missing crashes.

This also fixes the names for the models so they show up correctly
in the Django admin.