bug 1540705: remove .rst CoC and add CoC pointer text
This removes the .rst version of the CoC that we had since it doesn't
meet the Mozilla standard--only the .md version does.

This adds text to README and CONTRIBUTING to point to the CoC file
in the repository. The README is included in the Sphinx build which
ends up on ReadTheDocs.

This also adds a link to the footer of the website.