bug 1544919: Create API Token section
The SOCORRO_API_TOKEN needs to be set before running make shell,
to make it available for fetch_crash_data. Move it higher, so a new
developer is more likely to add it at the right time. List the two
permissions needed for crash data plus memory dumpos. Be more
explicit about the limits of crash data without PII.