bug 1544919: Remove incorrect "no ids" message
In a for/else statement, the else clause is called when the for loop
completes without breaking, but the code was written as if it was called
for an empty loop.

Since the "no crashids" case is handled in main (by printing usage
help), it doesn't need to be handled here as well.