bug 1545784: handle Deadline Exceeded in pubsub code
When our pubsub queue code got a "504 Deadline Exceeded", it would
let that bubble up all the way to the fetch_transform_and_save_app code
where it would get logged and then the processor would exit with
a 0 exit code. This causes the processor node to be dead and not dead
at the same time. It's hilarious.

This fixes the pubsub queue code such that it treats "504 Deadline
Exceeded" as if there were no crash ids to fetch and continues on
with its round-robin checking of queues. Seems like it's an ephemeral
error, so this is the right thing to do.