bug 1546433: add "es list" and fix config
This fixes how the es command creates a ConnectionContext such that it
now correctly uses the ConnectionContext defaults. This is nice because
if we ever change the ConnectionContext, we don't have to fix the
configuration in the es command, too. Fewer things will break.

This also adds a "list" subcommand so you can see what the current
state of things is.