bug 1546433: refactor Elasticsearch index delete code
This redoes all the code that has anything to do with deleting Elasticsearch
indices. Now all create/delete code is part of ConnectionContext.

This adds "delete" to the "es" script so we can use the es script exclusively
in the local dev environment.

This fixes test configuration so that we can test against the Elasticsearch
service without messing up the local dev environment.

This adds an es_conn fixture for testing with Elasticsearch.