bug 1542392: fix depcheck to tell us if it's using a safety api key
This fixes the depcheck command to tell us if it's using a safety api key.
Otherwise we have no idea.

This also fixes what it does if the key is bad. Before it would raise
an error with the text "b''". Now it tells you the api key is invalid.
Isn't that nice.