bug 1542388: fix archivescraper worker printing
The worker processes in archivescraper would print things to stdout
about what they were doing. The problem is that in the cronrun context,
this would go directly to stdout and not get formatted in mozlog format
and so it won't go through our logging pipeline well.

This fixes that by buffering that output and sending it back in the
result.