bug 1549717: fix module assert error
There are some crash reports where the module in the frame is empty. This was
causing the assertion to fail. Instead of asserting it's not empyt, we can
convert it to an "if" block and add the module to the frame information only if
it's there like we do with some other things. This works with the rest of the
processing pipeline and the webapp report view.