bug 1552973: Log missing boto files as warnings
When a file is missing when using the boto crashstorage, log as a
warning, using a dictionary of log format parameters, rather than as a
pre-formated error. Changing from error to warning avoids creating a
Sentry issue when logging, and changing to a format string combines into
one Sentry issue, and avoids string formatting if the log isn't picked
up by a handler.