bug 1553314: remove configuration from Rules
This adjusts the ProcessorPipeline to own configuration of the rules in the
pipeline. Configuration is now defined on that class and it passes in necessary
data through init args to the rules that require it.  In this way, rules are
just transformation classes and aren't configman-style classes.

This adds logging of the pipeline rules and their configuration. This will
make it a lot easier to notice and debug rules problems in the future.

Doing this also fixed configuration for the JitCrashCategorizeRule which was
totally busted and probably has been for some time. Now rule configuration is
namespaced which allows us to configure both the BreakpadStackwalkerRule2015
and JitCrashCategorizeRule.

This also removed all the confusing duplicated (and triplicated in some places)
configuration which was confusing and causing problems.