bug 1544246: add "fix_missing_module" pass to signature generation
This adds a "fix_missing_module" function that adds missing modules
for Rust 1.34 symbols. The fixed functions will then match items in
the prefix, sentinels, and irrelevant lists.

It's not comprehensive, but it catches a bunch of the symbols that
I saw looking at GeckoCrash crash reports. We can add other file/function
items to the lookup table as we discover them.