bug 1513346: Only capture and log PolyStorageError
PolyStorageError holds exceptions raised by the storage backends. There
are no other exceptions in our code or key external libraries that
implement Sequence.

This simplifies the error handler to specifically handle
PolyStorageError rather than theoretical exceptions implementing
Sequence. The behaviour is the same, except that the message
"error in processing or saving crash" is logged once per processed crash
instead of once per failing storage.

Other exceptions are not handled here (not logged in Sentry, no log that
crash storage failed), and are instead propagated up to be handled
later, which is a change in behaviour.