bug 1473068: Add more OIDC_EXEMPT_URLs
Add more URLs that should be exempt from session refresh, because they
are used to load tabs in a signature report.

There are two more URLs for aggregations and graphs that should be
included, but these include a parameter (aggregation and field) in
the URL, so they can't be exactly matched in the middleware.