bug 1552898: implement manual featured versions
This implements manually-curated featured versions. They exist in the
product_details/ directory. People can edit and create pull requests using the
GitHub interface. Using GitHub for process flow significantly reduces the
complexity of the implementation. This is probably a good enough first pass
and we can adjust that as needs necessitate.

If a file exists, Crash Stats will use that to set the featured versions which
show up on the product home page and in the "featured version" part of the
current versions drop down in the menu. If a file does not exist or can't be
retrieved, Crash Stats will calculate the featured versions based on the
existing crash report data. Featured versions are cached for an hour after
which they're figured again, so if anything goes wrong, we just have to wait
a bit.