bug 1564551: Drop invalid JSON handler
The "no JSON" ValueError handler worked on Python 2.7, but is broken in
Python 3 (which removes implicit string conversions). It appears this
exception would be raised by the requests library, when converting a
response to JSON.

There is no test coverage, so this wasn't detected during the Python 3
conversion. There are no "no JSON" tracebacks from stage or production,
so this handler appears to be written for a old an no longer relevant
problem.

This commit removes the handler. If the "no JSON" issue appears again,
the handler can be re-added, along with tests.