bug 1564551: Check items in MultipleStringField
Rewrite MultipleStringField to run validators against each item in the
field. This is different than most Django fields, which validate the
entire value (in this case, a list of strings).

Add ProhibitNullCharactersValidator(), to avoid accepting strings that
will be rejected by the database backends.

Continue to avoid checking that the value is in self.choices, which is
the validation strategy of the base class TypedMultipleChoiceField.