bug 1552709: Escape surrogates on report index
A processed dump may contain a surrogate character in the range
U+DC80 to U+DCFF, used to represent non-decodable bytes on POSIX systems
(see PEP 0383). By default, these can not be encoded in UTF-8, and an
error handler is needed. This escapes the characters, so (for example)
U+DF03 becomes the string "\udf03" in the UTF-8 encoding.