bug 1503344: start putting data in java_stack_trace_raw
We have this problem where I named the field with _full at the end not
knowing about the whole _full thing our Elasticsearch
crashstorage/supersearch thing does. We periodically get errors from
that poor choice.

This does the first step in renaming the field to _raw--it additionally
puts the data in that field and deprecates the original one. It also
adds a note to remove the old field in January 2020.