bug 1559223, 1542964: add modules_in_stack to processed crash
This adds a new "modules_in_stack" field which is a semicolon-delimited
set of "module/debugid" strings for modules that are in the stack of the
crashing thread.

The value is indexed using the semicolon_analyzer, so each string ends
up as a single term. This allows us to use supersearch "startswith" for
finding crash reports with a specified module in the stack (e.g.
"libfenix.so") and "has term" for finding crash reports with a specified
"module/debugid" in the crashing thread of the stack. The latter helps
us reprocess crash reports for which we just uploaded symbols.