bug 1433148: switch from boto to boto3
This switches the code from boto to boto3. Because that effectively
required rewriting some of the code, I ended up rewriting all of it.

The S3Connection class is based on the one we have in Antenna with some
changes:

* uses configman instead of everett
* has a load_file method
* nixes the bucket check code
* has better tests

Because S3Connection supports the local dev environment and server
environments and a variety of different AWS authentication methods, this
removes all the other connection context variants since we don't need
them anymore.

This redoes the tests for crash_data and crashstorage.

This adds docstrings for methods along with notes and what the methods
might raise.

Since S3Connection handles retrying, this removes the retrying in the
crashstorage code.

This changes some of the configuration. We no longer have AWS_HOST,
AWS_PORT, and AWS_SECURE. Those have been replaced with
AWS_ENDPOINT_URL. This only affects the local dev environment.

This adds ResourceWarning to the warnings to ignore in pytest. boto3
raises a ResourceWarning since it doesn't explicitly close connections.

This gets rid of moto in favor of tests hitting localstack-s3 service
directly.