bug 1433148: remove addressing_style setting
In 2018, they changed botocore so that it defaults to "auto" which will
start with "virtual" and if that doesn't work, switch to "path". We no
longer have periods in the bucket names, so the only situation where we
need "path" is the local dev environment.

Given that, we don't need to set the addressing_style to "path" plus the
comment isn't correct anymore.