bug 1603236: fix unknown property indexing failure
Let's say we have two crash reports that have a field that isn't in
super search fields. The first crash report has a field value that's a
big string. The second one has a nested value. When the first is
indexed, Elasticsearch will infer that field to be a string. When the
second gets indexed, it'll raise an "unknown property" error and not
index the crash report.

This fixes that situation by adding code that recognizes what just
happened, drops the field from the document, and retries indexing.
This adds a test, too. This also adds metrics so we can see how often
these indexing errors happen.